Switched coordinatesystem for model in/output
Switched coord. system for the model input and output to be similar to
Sofi3D. Unlike in the code where z is the fast dimension, the in/output
uses y (depth) as fast and z as slow dimension.
That means by using for example ximage(2D) to visualisze the model you will see NZ x-y planes.